Add files via upload
Write a Java program to create a table Employee in the database using a statement interface and check if the table has been created or not.The add few records from mysql command line client and execute a simple SELECT query to retrieve data from a table.. [Hint:The "Employee" table is designed to include an "eid" column as the primary key, which will auto-increment its values. It also contains "name" and "address" columns of type varchar, and a "salary" column of type int. All columns in the table are set to be non-nullable.]